pyproject.toml
[project]
name = "pdscan"
version = "0.1.0"
description = "Scan your data stores for unencrypted personal data (PII)"
readme = "README.md"
requires-python = ">=3.7"
license = {text = "MIT"}

authors = [
  { name = "Your Name", email = "your.email@example.com" }
]

dependencies = [
  "click>=8.0.0",
  "boto3>=1.26.0",
  "elasticsearch>=8.0.0",
  "pymongo>=4.0.0",
  "redis>=4.0.0",
  "psycopg2-binary>=2.9.0",
  "mysql-connector-python>=8.0.0",
  "sqlalchemy>=2.0.0",
  "pandas>=2.0.0",
  "openpyxl>=3.1.0",
  "python-magic-bin>=0.4.14",
  "pyyaml>=6.0",
  "cryptography>=3.4.8",
  "jinja2>=3.0.0",
  "reportlab>=3.6.0",
  "fastapi>=0.68.0",
  "uvicorn>=0.15.0",
  "httpx>=0.24.0",
  "pydantic>=1.10.0",       
  "opensearch-py>=2.0.0",
  "cx_Oracle>=8.0.0",
  "psutil>=5.9.0",
  "rich>=13.0.0",
]

[project.scripts]
pdscan = "pdscan.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
